agents/Youtube Trend analysis/requirements.txt
streamlit>=1.28.0
crewai>=0.28.0
crewai-tools>=0.1.0
python-dotenv>=1.0.0
langchain-openai<0.2.0,>=0.1.7
langchain<0.2.0,>=0.1.10
openai<2.0.0,>=1.13.3
youtube-transcript-api>=0.6.0
requests>=2.31.0
pyyaml>=6.0
tqdm>=4.65.0
